Dockerization Proposal
