+++ access/Dockerfile
@@ -1,0 +1,5 @@
+FROM golang:onbuild
+
+MAINTAINER Laura Chaparro <lachaparrog@unal.edu.co>
+
+EXPOSE 4025
