+++ start.sh
@@ -1,0 +1,2 @@
+sudo docker-compose build
+sudo docker-compose up
